myjson and majson should work, ojson has flows, pgjsonb still to do
